+++ src/main/java/io/github/jamalam360/jamlib/JamLib.java
@@ -27,18 +27,14 @@ package io.github.jamalam360.jamlib;
 import io.github.jamalam360.jamlib.tick.TickScheduling;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerTickEvents;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
+import io.github.jamalam360.jamlib.log.JamLibLogger;
 
 public class JamLib implements ModInitializer {
-    public static Logger getLogger(String subName) {
-        return LogManager.getLogger("JamLib/" + subName);
-    }
+    public static final JamLibLogger LOGGER = JamLibLogger.getLogger("jamlib");
 
     @Override
     public void onInitialize() {
         ServerTickEvents.END_WORLD_TICK.register(TickScheduling::onEndTickServer);
-
-        getLogger("JamLibInit").info("JamLib has been initialized");
+        LOGGER.logInitialize();
     }
 }