+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 ## Changelog
 
-### Fixes
+### Features
 
-- More 1.19 related config fixes.
+- Add a logging helper.
+- Revamp the registry system to be annotation based.
+
+### Fixes
 
 Closed issues: None.
 
-[Full Changelog](https://github.com/JamCoreModding/JamLib/compare/0.0.7...0.0.8)
+[Full Changelog](https://github.com/JamCoreModding/JamLib/compare/0.0.8...0.1.0)